backend/pyproject.toml
[project]
name = "ontosys-backend"
version = "0.1.0"
description = "Event-Storming to Neo4j Graph Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "neo4j>=5.17.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "tiktoken>=0.5.2",
    "python-multipart>=0.0.6",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "redis>=5.0.1",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.1",
    "rapidfuzz>=3.6.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.26.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.26.0",
    "pytest-asyncio>=0.23.0",
]



